backend/requirements.txt
# Agent Mitra - Python Backend Dependencies
# FastAPI and Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Redis
redis==5.2.0
hiredis==2.3.2

# Environment & Configuration
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
psutil==6.0.0

# AI/ML
openai==1.51.0

# Feature Flag Management - FeatureHub
# Install FeatureHub SDK: pip install featurehub-sdk
# Using httpx for REST API integration (can switch to SDK later)
# featurehub-sdk==1.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Development
black==24.10.0
flake8==7.1.1
mypy==1.11.2

